[Update] cls_moto.py

+++ DEVinCar/classes/cls_moto.py
@@ -36,7 +36,7 @@ class Moto(Veiculo):
 
         placa = str(input("Digite a placa da moto: "))
         cor_nova = str(input(f"Digite a nova cor: "))
-        valor_novo = int(input(f"Digite o valor novo R$: "))
+        valor_novo = float(input(f"Digite o valor novo R$: "))
         
         for x in lista_motos:
